- Add tests for ResponseTimeMiddleware
